chore: Reorganize templates folder
