+++ src/opex_dashboard/templates/availability-query.kusto
