+++ src/opex_dashboard/templates/response-codes-query.kusto
