+++ src/opex_dashboard/templates/response-time-query.kusto
