+++ src/opex_dashboard/templates/template-part.json
