+++ src/opex_dashboard/templates/template.json
